src_collect/requirements.txt
mysql
python-dotenv
requests